feat: overhaul build system
No longer depend on my own package, but instead go directly to the
various sources. There's a real nightmare in the different versions, and
although the "official" sokol `dcimgui` works really well for the sample
setup, it becomes hell when trying to link with cimplot or docking. For
this reason, we maintain our own sources.

Builds natively as before.

WASM not yet working.